+++ libraries/pod/src/lib.rs
@@ -2,8 +2,8 @@
 
 pub mod bytemuck;
 pub mod error;
+pub mod option;
 pub mod optional_keys;
-pub mod pod_option;
 pub mod primitives;
 pub mod slice;
